Reverse words
Hi, here's your problem today. This problem was recently asked by Apple:

Given a list of words in a string, reverse the words in-place (ie don't create a new string and reverse the words). Since python strings are not mutable, you can assume the input will be a mutable sequence (like list).


Without extra space, a string reversal problem like this usually would require some combination of in-place string reversal. If we reversed the whole string, you'll see that the order of words are reversed, but each word itself are also reversed. So if we run through the whole string and reversed each word so it's in its original order, we'll have our answer. We can tell the beginning and end of each word by iterating through the string until we find a space, in which we can reverse only that part of the string.

With this general solution, there are some edge cases to take care of, specifically related to whitespaces. We need to take care of the case where there is more than 1 space in between words, using a while loop when we see a space until we find the next character that is not a space.


The time complexity is O(n) for doing at most n swaps (0.5n swaps for reversing the whole string, and another 0.5n swaps for reversing each word), and the space complexity is O(1).